% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{add_project_issue}
\alias{add_project_issue}
\title{Add Issue to Github Project}
\usage{
add_project_issue(gq, project_id, issue_id, verbose = FALSE)
}
\arguments{
\item{gq}{GraphQL connection object from `get_gq()`}

\item{project_id}{Project ID}

\item{issue_id}{Issue ID}

\item{verbose}{Print GraphQL query; default=FALSE}
}
\value{
ID of new item in project
}
\description{
Add Issue to Github Project
}
\examples{
\dontrun{
 token <- Sys.getenv("GITHUB_TOKEN")
 gq <- get_gq(token)
 prj <- get_project_id(gq, owner = "MarineSensitivities", num = 1)
 #TODO: get issues from repository
}
}
