% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_gq}
\alias{get_gq}
\title{Get GraphQL connection}
\usage{
get_gq(token, url = "https://api.github.com/graphql")
}
\arguments{
\item{token}{Github personal access token. See \href{https://docs.github.com/en/graphql/guides/forming-calls-with-graphql#authenticating-with-graphql}{Authenticating with GraphQL}.}

\item{url}{GraphQL API URL. Default is \code{https://api.github.com/graphql}}
}
\value{
\link[ghql]{GraphqlClient} object
}
\description{
Get GraphQL connection to Github. Refresh schema before returning.
}
\examples{
\dontrun{
  token <- Sys.getenv("GITHUB_TOKEN")
  gq <- get_gq(token)
  gq
}
}
