% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_owner_repos}
\alias{get_owner_repos}
\title{Get repositories for a Github user or organization}
\usage{
get_owner_repos(owner = "MarineSensitivities")
}
\arguments{
\item{owner}{Github user or organization name}
}
\value{
Character vector of repository names
}
\description{
Get repositories for a Github user or organization
}
\examples{
\dontrun{
  get_owner_repos("MarineSensitivities")
}
}
