% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_project_id}
\alias{get_project_id}
\title{Get Github Project ID}
\usage{
get_project_id(gq, owner = "MarineSensitivities", project_num = 1)
}
\arguments{
\item{gq}{GraphQL connection object from `get_gq()`}

\item{owner}{Github user or organization name containing the Project}

\item{project_num}{Project number}
}
\value{
Project ID
}
\description{
Needed to operate on GraphQL.
}
\examples{
\dontrun{
  token <- Sys.getenv("GITHUB_TOKEN")
  gq <- get_gq(token)
  get_project_id(gq, owner = "MarineSensitivities", num = 1)
}
}
