% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_project_issues}
\alias{get_project_issues}
\title{Get Project Issues}
\usage{
get_project_issues(gq, project_id)
}
\arguments{
\item{gq}{GraphQL connection object from `get_gq()`}

\item{project_id}{Project ID from `get_project_id()`}
}
\value{
Character vector of issue node IDs, with titles as names
}
\description{
Get Project Issues
}
\examples{
\dontrun{
 token <- Sys.getenv("GITHUB_TOKEN")
 gq <- get_gq(token)
 prj <- get_project_id(gq, owner = "MarineSensitivities", num = 1)
 get_project_issues(gq, prj)
}
}
