% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_repo_issues}
\alias{get_repo_issues}
\title{Get Issues from Github repository}
\usage{
get_repo_issues(owner = "MarineSensitivities", repo = "server")
}
\arguments{
\item{owner}{Github user or organization name containing issues}

\item{repo}{Github repository name}
}
\value{
Character vector of issue node IDs
}
\description{
Get Issues from Github repository
}
\examples{
\dontrun{
 get_repo_issues("MarineSensitivities", "server")
}
}
