% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{glueb}
\alias{glueb}
\title{Glue with brackets}
\usage{
glueb(x, ...)
}
\arguments{
\item{...}{Arguments passed to \link[glue]{glue}}
}
\value{
character vector
}
\description{
Since GraphQL statements heavily use curly braces `{}`,
this function is a wrapper around \link[glue]{glue} to make it easier to
write GraphQL statements.
}
\examples{
 glueb("Hello {name}", name = "world")
}
